[General Setup] - Import PX4 Autopilot Repo as a Submodule
Problem

We need to pull in the base firmware we will be using for our flight
controller. We have chosen to use `PX4` instead of `Ardupilot` for this
task.

Solution

Import `PX4/PX4-Autopilot` repo as a submodule in the `/src` directory.
Imported here rather than in either `/dev` or `/prod` as the majority of
our code will likely be implemented on top of this repo, not within it.

Testing

None required.

Issue: https://github.com/ameall/DRIFT_FW/issues/1