Issue 9: Receive Camera Frames (#16)
Problem

In a previous task, we connected and configured our Raspberry Pi Global
Shutter Camera to verify that we were able to receive data from our
camera. Now that we know the camera is working, we need to be able to
get image frames from the camera in a program-usable format, so that
our `CV` algorithm can run on the data that the camera sees.

Solution

Design a base camera application to verify that we can properly interact
with and receive images from the camera.

Testing

On-device runs; a test image is saved as "image1.jpg".

Issue: https://github.com/ameall/drift-fw/issues/9