Issue 17 - CV Model Integration (#17)
Problem

Now that we've created the camera app and verified it's functionality,
we need to link it to the CV model we will be using for processing the
camera frames.

Solution

Implemented functions to convert image frames received from the camera
into a format that is usable by our CV model and test that images are
being passed into the model correctly.

Testing

For an unknwon reason, the cv::imshow() function causes the application
to break if inserted into the camera request callback function. Not 100%
sure about functionality, but have made reasonable assumptions and
tested at various points to be as confident as possible.

Would normally deep-dive this, but this code is needed asap so will have
to come back to it later.