Issue 20: Inter-App Communication (#20)
Problem

Now that we have developed teh camera app and integrated it with the CV
model, we must define a method for communicating results from the camera
app to the flight control app.

Solution

Implemented the client and server side of a UNIX socket to allow for
communication between the two processes, as well as defined a message
structure to allow for all necessary data to be communicated between the
camera app and the flight control app.

Testing

Tested UNIX socket directly

Issue: https://github.com/ameall/drift-fw/issues/20