Issue 31: Client-Side Camera Simulator (#31)
Problem

We went out for a test run a few days ago. As part of this run, we
tested using a camera simulator on the client-side of the socket, to
more closely replicate a black-box camera application.

Solution

Push the code that was written on site.

Testing

Flight tests verified code to be working

Issue: https://github.com/ameall/drift-fw/issues/31