Issue 34: BUG: Model Reallocation On Every Frame (#34) (#49)
Problem

It was recently discovered that the camera app allocates and frees the
CV model on every instance of the callback function, which results in
unnecessary memory operations. This also have the unintended consequence
of disconnecting and reconnecting the UNIX socket for every frame.

Task

Modified the camera app CV model/frame request callback function to not
reallocate the model on every frame.

Testing

On-drone testing

Issue: https://github.com/ameall/drift-fw/issues/34